refactor: removed reexport of virtio modules
This is done to make it clear where the imports are comming from.
