+++ src/vmm/src/devices/bus.rs
@@ -56,7 +56,7 @@ use event_manager::{EventOps, Events, MutEventSubscriber};
 use super::legacy::RTCDevice;
 use super::legacy::{I8042Device, SerialDevice};
 use super::pseudo::BootTimer;
-use super::virtio::MmioTransport;
+use super::virtio::mmio::MmioTransport;
 
 #[derive(Debug)]
 pub enum BusDevice {
