+++ src/vmm/src/devices/virtio/balloon/event_handler.rs
@@ -6,9 +6,10 @@ use std::os::unix::io::AsRawFd;
 use event_manager::{EventOps, Events, MutEventSubscriber};
 use utils::epoll::EventSet;
 
+use super::{DEFLATE_INDEX, INFLATE_INDEX, STATS_INDEX};
 use crate::devices::report_balloon_event_fail;
 use crate::devices::virtio::balloon::device::Balloon;
-use crate::devices::virtio::{VirtioDevice, DEFLATE_INDEX, INFLATE_INDEX, STATS_INDEX};
+use crate::devices::virtio::device::VirtioDevice;
 use crate::logger::{debug, error, warn};
 
 impl Balloon {
