+++ src/vmm/src/devices/virtio/balloon/test_utils.rs
@@ -7,12 +7,13 @@ use std::u32;
 
 use crate::devices::virtio::test_utils::VirtQueue;
 #[cfg(test)]
-use crate::devices::virtio::{
-    balloon::BALLOON_NUM_QUEUES, Balloon, IrqType, DEFLATE_INDEX, INFLATE_INDEX, STATS_INDEX,
-};
+use crate::devices::virtio::{balloon::Balloon, balloon::BALLOON_NUM_QUEUES};
 
 #[cfg(test)]
 pub fn invoke_handler_for_queue_event(b: &mut Balloon, queue_index: usize) {
+    use crate::devices::virtio::balloon::{DEFLATE_INDEX, INFLATE_INDEX, STATS_INDEX};
+    use crate::devices::virtio::device::IrqType;
+
     assert!(queue_index < BALLOON_NUM_QUEUES);
     // Trigger the queue event.
     b.queue_evts[queue_index].write(1).unwrap();
