+++ src/vmm/src/devices/virtio/block/persist.rs
@@ -11,12 +11,14 @@ use utils::eventfd::EventFd;
 use versionize::{VersionMap, Versionize, VersionizeError, VersionizeResult};
 use versionize_derive::Versionize;
 
+use super::device::DiskProperties;
 use super::*;
-use crate::devices::virtio::block::device::{DiskProperties, FileEngineType};
-use crate::devices::virtio::block_metrics::BlockMetricsPerDevice;
+use crate::devices::virtio::block::block_metrics::BlockMetricsPerDevice;
+use crate::devices::virtio::block::device::FileEngineType;
+use crate::devices::virtio::device::{DeviceState, IrqTrigger};
 use crate::devices::virtio::gen::virtio_blk::VIRTIO_BLK_F_RO;
 use crate::devices::virtio::persist::VirtioDeviceState;
-use crate::devices::virtio::{DeviceState, IrqTrigger, FIRECRACKER_MAX_QUEUE_SIZE, TYPE_BLOCK};
+use crate::devices::virtio::TYPE_BLOCK;
 use crate::logger::warn;
 use crate::rate_limiter::persist::RateLimiterState;
 use crate::rate_limiter::RateLimiter;
