+++ src/vmm/src/devices/virtio/block/test_utils.rs
@@ -12,14 +12,15 @@ use utils::kernel_version::{min_kernel_version_for_io_uring, KernelVersion};
 use utils::tempfile::TempFile;
 
 use super::device::VirtioBlockConfig;
+use super::RequestHeader;
 use crate::devices::virtio::block::device::FileEngineType;
 #[cfg(test)]
 use crate::devices::virtio::block::io::FileEngine;
-use crate::devices::virtio::queue::{VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE};
-use crate::devices::virtio::test_utils::{VirtQueue, VirtqDesc};
+use crate::devices::virtio::block::{CacheType, VirtioBlock};
 #[cfg(test)]
-use crate::devices::virtio::IrqType;
-use crate::devices::virtio::{CacheType, Queue, RequestHeader, VirtioBlock};
+use crate::devices::virtio::device::IrqType;
+use crate::devices::virtio::queue::{Queue, VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE};
+use crate::devices::virtio::test_utils::{VirtQueue, VirtqDesc};
 use crate::rate_limiter::RateLimiter;
 use crate::vmm_config::{RateLimiterConfig, TokenBucketConfig};
 use crate::vstate::memory::{Bytes, GuestAddress};
