+++ src/vmm/src/devices/virtio/vsock/mod.rs
@@ -33,7 +33,7 @@ use crate::devices::virtio::persist::PersistError as VirtioStateError;
 use crate::vstate::memory::GuestMemoryMmap;
 
 mod defs {
-    use crate::devices::virtio::FIRECRACKER_MAX_QUEUE_SIZE;
+    use crate::devices::virtio::queue::FIRECRACKER_MAX_QUEUE_SIZE;
 
     /// Device ID used in MMIO device identification.
     /// Because Vsock is unique per-vm, this ID can be hardcoded.
