+++ src/vmm/src/devices/virtio/vsock/test_utils.rs
@@ -9,12 +9,13 @@ use std::os::unix::io::{AsRawFd, RawFd};
 use utils::epoll::EventSet;
 use utils::eventfd::EventFd;
 
+use crate::devices::virtio::device::VirtioDevice;
+use crate::devices::virtio::queue::{VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE};
 use crate::devices::virtio::test_utils::{single_region_mem, VirtQueue as GuestQ};
 use crate::devices::virtio::vsock::device::{RXQ_INDEX, TXQ_INDEX};
 use crate::devices::virtio::vsock::packet::{VsockPacket, VSOCK_PKT_HDR_SIZE};
-use crate::devices::virtio::{
-    VirtioDevice, Vsock, VsockBackend, VsockChannel, VsockEpollListener, VsockError,
-    VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE,
+use crate::devices::virtio::vsock::{
+    Vsock, VsockBackend, VsockChannel, VsockEpollListener, VsockError,
 };
 use crate::vstate::memory::{GuestAddress, GuestMemoryMmap};
