+++ src/vmm/src/lib.rs
@@ -129,11 +129,12 @@ use crate::cpu_config::templates::CpuConfiguration;
 use crate::device_manager::legacy::PortIODeviceManager;
 use crate::device_manager::mmio::MMIODeviceManager;
 use crate::devices::legacy::{IER_RDA_BIT, IER_RDA_OFFSET};
-use crate::devices::virtio::balloon::BalloonError;
-use crate::devices::virtio::{
-    Balloon, BalloonConfig, BalloonStats, Net, VirtioBlock, BALLOON_DEV_ID, TYPE_BALLOON,
-    TYPE_BLOCK, TYPE_NET,
+use crate::devices::virtio::balloon::{
+    Balloon, BalloonConfig, BalloonError, BalloonStats, BALLOON_DEV_ID,
 };
+use crate::devices::virtio::block::VirtioBlock;
+use crate::devices::virtio::net::Net;
+use crate::devices::virtio::{TYPE_BALLOON, TYPE_BLOCK, TYPE_NET};
 use crate::logger::{error, info, warn, MetricsError, METRICS};
 use crate::persist::{MicrovmState, MicrovmStateError, VmInfo};
 use crate::rate_limiter::BucketUpdate;
