+++ src/vmm/src/vmm_config/net.rs
@@ -9,8 +9,7 @@ use serde::{Deserialize, Serialize};
 use utils::net::mac::MacAddr;
 
 use super::RateLimiterConfig;
-use crate::devices::virtio::net::TapError;
-use crate::devices::virtio::Net;
+use crate::devices::virtio::net::{Net, TapError};
 use crate::VmmError;
 
 /// This struct represents the strongly typed equivalent of the json body from net iface
